Move license utilities into their own file
