+++ app/src/main/java/fr/free/nrw/commons/upload/license/MediaLicensePresenter.kt
@@ -1,9 +1,9 @@
 package fr.free.nrw.commons.upload.license
 
-import fr.free.nrw.commons.utils.Utils
 import fr.free.nrw.commons.kvstore.JsonKvStore
 import fr.free.nrw.commons.repository.UploadRepository
 import fr.free.nrw.commons.settings.Prefs
+import fr.free.nrw.commons.utils.Licenses
 import timber.log.Timber
 import java.lang.reflect.Method
 import java.lang.reflect.Proxy
@@ -39,7 +39,7 @@ class MediaLicensePresenter @Inject constructor(
             Prefs.Licenses.CC_BY_SA_4
         ) //CC_BY_SA_4 is the default one used by the commons web app
         try { //I have to make sure that the stored default license was not one of the deprecated one's
-            Utils.licenseNameFor(selectedLicense)
+            Licenses.licenseNameFor(selectedLicense)
         } catch (exception: IllegalStateException) {
             Timber.e(exception)
             selectedLicense = Prefs.Licenses.CC_BY_SA_4
