+++ app/src/main/java/fr/free/nrw/commons/utils/Licenses.java
@@ -1,0 +1,52 @@
+package fr.free.nrw.commons.utils;
+
+import androidx.annotation.NonNull;
+import fr.free.nrw.commons.R;
+import fr.free.nrw.commons.settings.Prefs;
+
+public final class Licenses {
+
+    /**
+     * Generates licence name with given ID
+     * @param license License ID
+     * @return Name of license
+     */
+    public static int licenseNameFor(String license) {
+        switch (license) {
+            case Prefs.Licenses.CC_BY_3:
+                return R.string.license_name_cc_by;
+            case Prefs.Licenses.CC_BY_4:
+                return R.string.license_name_cc_by_four;
+            case Prefs.Licenses.CC_BY_SA_3:
+                return R.string.license_name_cc_by_sa;
+            case Prefs.Licenses.CC_BY_SA_4:
+                return R.string.license_name_cc_by_sa_four;
+            case Prefs.Licenses.CC0:
+                return R.string.license_name_cc0;
+        }
+        throw new IllegalStateException("Unrecognized license value: " + license);
+    }
+
+    /**
+     * Generates license url with given ID
+     * @param license License ID
+     * @return Url of license
+     */
+    @NonNull
+    public static String licenseUrlFor(String license) {
+        switch (license) {
+            case Prefs.Licenses.CC_BY_3:
+                return "https://creativecommons.org/licenses/by/3.0/";
+            case Prefs.Licenses.CC_BY_4:
+                return "https://creativecommons.org/licenses/by/4.0/";
+            case Prefs.Licenses.CC_BY_SA_3:
+                return "https://creativecommons.org/licenses/by-sa/3.0/";
+            case Prefs.Licenses.CC_BY_SA_4:
+                return "https://creativecommons.org/licenses/by-sa/4.0/";
+            case Prefs.Licenses.CC0:
+                return "https://creativecommons.org/publicdomain/zero/1.0/";
+            default:
+                throw new IllegalStateException("Unrecognized license value: " + license);
+        }
+    }
+}
