+++ app/src/main/java/fr/free/nrw/commons/utils/Utils.java
@@ -11,7 +11,6 @@ import android.text.style.UnderlineSpan;
 import android.view.View;
 import android.widget.TextView;
 
-import androidx.annotation.NonNull;
 import androidx.browser.customtabs.CustomTabColorSchemeParams;
 import androidx.browser.customtabs.CustomTabsIntent;
 import androidx.core.content.ContextCompat;
@@ -25,54 +24,9 @@ import java.util.Locale;
 import java.util.regex.Pattern;
 
 import fr.free.nrw.commons.location.LatLng;
-import fr.free.nrw.commons.settings.Prefs;
 import timber.log.Timber;
 
 public class Utils {
-
-    /**
-     * Generates licence name with given ID
-     * @param license License ID
-     * @return Name of license
-     */
-    public static int licenseNameFor(String license) {
-        switch (license) {
-            case Prefs.Licenses.CC_BY_3:
-                return R.string.license_name_cc_by;
-            case Prefs.Licenses.CC_BY_4:
-                return R.string.license_name_cc_by_four;
-            case Prefs.Licenses.CC_BY_SA_3:
-                return R.string.license_name_cc_by_sa;
-            case Prefs.Licenses.CC_BY_SA_4:
-                return R.string.license_name_cc_by_sa_four;
-            case Prefs.Licenses.CC0:
-                return R.string.license_name_cc0;
-        }
-        throw new IllegalStateException("Unrecognized license value: " + license);
-    }
-
-    /**
-     * Generates license url with given ID
-     * @param license License ID
-     * @return Url of license
-     */
-    @NonNull
-    public static String licenseUrlFor(String license) {
-        switch (license) {
-            case Prefs.Licenses.CC_BY_3:
-                return "https://creativecommons.org/licenses/by/3.0/";
-            case Prefs.Licenses.CC_BY_4:
-                return "https://creativecommons.org/licenses/by/4.0/";
-            case Prefs.Licenses.CC_BY_SA_3:
-                return "https://creativecommons.org/licenses/by-sa/3.0/";
-            case Prefs.Licenses.CC_BY_SA_4:
-                return "https://creativecommons.org/licenses/by-sa/4.0/";
-            case Prefs.Licenses.CC0:
-                return "https://creativecommons.org/publicdomain/zero/1.0/";
-            default:
-                throw new IllegalStateException("Unrecognized license value: " + license);
-        }
-    }
 
     /**
      * Adds extension to filename. Converts to .jpg if system provides .jpeg, adds .jpg if no extension detected
