+++ app/src/test/kotlin/fr/free/nrw/commons/upload/MediaLicensePresenterTest.kt
@@ -6,6 +6,7 @@ import fr.free.nrw.commons.kvstore.JsonKvStore
 import fr.free.nrw.commons.repository.UploadRepository
 import fr.free.nrw.commons.upload.license.MediaLicenseContract
 import fr.free.nrw.commons.upload.license.MediaLicensePresenter
+import fr.free.nrw.commons.utils.Licenses
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
@@ -50,7 +51,7 @@ class MediaLicensePresenterTest {
         MockitoAnnotations.openMocks(this)
         mediaLicensePresenter.onAttachView(view)
         mockedUtil = Mockito.mockStatic(Utils::class.java)
-        `when`(Utils.licenseNameFor(ArgumentMatchers.anyString())).thenReturn(1)
+        `when`(Licenses.licenseNameFor(ArgumentMatchers.anyString())).thenReturn(1)
     }
 
     @After
